Change Type.value() and DClass.value() to return an int, not a short, so that the entire range of types/classes as well as an error value can be returned. Update all of the callers.
git-svn-id: http://svn.code.sf.net/p/dnsjava/code/trunk@1046 c76caeb1-94fd-44dd-870f-0c9d92034fc1

+++ lookup.java
@@ -29,7 +29,7 @@ printAnswer(String name, Lookup lookup) {
 
 public static void
 main(String [] args) throws Exception {
-	short type = Type.A;
+	int type = Type.A;
 	int start = 0;
 	if (args.length > 2 && args[0].equals("-t")) {
 		type = Type.value(args[1]);
@@ -38,7 +38,7 @@ main(String [] args) throws Exception {
 		start = 2;
 	}
 	for (int i = start; i < args.length; i++) {
-		Lookup l = new Lookup(args[i], type);
+		Lookup l = new Lookup(args[i], (short) type);
 		l.run();
 		printAnswer(args[i], l);
 	}
